Title: 최소 에너지 비용, Time: 310ms, Memory: 23MB, Status: Passed - Codetree

+++ 250726/최소 에너지 비용/minimum-energy-cost.java
@@ -3,14 +3,14 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
-        int[] dist = new int[n - 1];
-        int[] cost = new int[n];
+        long[] dist = new long[n - 1];
+        long[] cost = new long[n];
         for (int i = 0; i < n - 1; i++)
             dist[i] = sc.nextInt();
         for (int i = 0; i < n; i++)
             cost[i] = sc.nextInt();
 
-        int[] lowCost = new int[n];
+        long[] lowCost = new long[n];
         lowCost[0] = cost[0];
         for (int i = 1; i < n; i++)
             lowCost[i] = Math.min(lowCost[i - 1], cost[i]);
